♻️  F12: BasketBody > Separated content if empty or not in two separate components
