+++ src/components/pages/order/Main/Basket/BasketBody.js
@@ -1,49 +1,15 @@
 import React, { useContext } from "react";
 import styled from "styled-components";
-import { theme } from "../../../../../theme";
-import BasketCard from "../../../../reusable/BasketCard";
 import OrderContext from "../../../../../context/OrderContext";
-import { formatPrice } from "../../../../../utils/maths";
-import { DEFAULT_IMAGE_SOURCE } from "../../../../../js/enum";
+import EmptyBasket from "./EmptyBasket";
+import BasketCardList from "./BasketCardList";
 
 export default function BasketBody() {
-  const {
-    basket,
-    menu,
-    handleRemoveFromBasket,
-    isAdmin,
-    verifyIfCardIsSelected,
-    selectProductToEdit,
-  } = useContext(OrderContext);
+  const { basket } = useContext(OrderContext);
 
   return (
     <BasketBodyStyled>
-      {basket.length === 0 ? (
-        <div className="empty">Votre commande est vide.</div>
-      ) : (
-        <div className="produits">
-          {basket.map((product) => {
-            const productInfo = menu.find((item) => item.id === product.id);
-            return (
-              <BasketCard
-                key={product.id}
-                imageSource={
-                  productInfo.imageSource
-                    ? productInfo.imageSource
-                    : DEFAULT_IMAGE_SOURCE
-                }
-                title={productInfo.title}
-                price={formatPrice(productInfo.price)}
-                quantity={product.quantity}
-                onDelete={() => handleRemoveFromBasket(product.id)}
-                isHoverable={isAdmin}
-                isSelected={verifyIfCardIsSelected(product.id)}
-                onClick={() => selectProductToEdit(product.id)}
-              />
-            );
-          })}
-        </div>
-      )}
+      {basket.length === 0 ? <EmptyBasket /> : <BasketCardList />}
     </BasketBodyStyled>
   );
 }
@@ -54,27 +20,4 @@ const BasketBodyStyled = styled.div`
   align-items: center;
   overflow: hidden;
   z-index: 2;
-
-  .empty {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: ${theme.font.sizes.P4};
-    font-family: ${theme.font.families.stylish};
-    color: ${theme.colors.greyBlue};
-  }
-
-  .produits {
-    padding: 10px 0px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    overflow-y: scroll;
-
-    ::-webkit-scrollbar {
-      display: none;
-    }
-  }
 `;
